+++ LyraEditor/Source/EditorLayer.h
@@ -352,7 +352,7 @@ namespace Lyra
 				ImGui::Text((std::string("Mesh count: ") + std::to_string(meshes.size())).c_str());
 				for (auto& mesh : meshes)
 				{
-					ImGui::PushID(mesh->GetHash());
+					ImGui::PushID((int)mesh->GetHash());
 
 					if (ImGui::CollapsingHeader(mesh->GetName().data()))
 					{
